// src/Types/google.maps.d.ts
export interface GeocoderAddressComponent {
  long_name: string;
  short_name: string;
  types: string[];
}
// ...
export interface MapsEventListener {}
// ...
export class LatLng {
  /**
   * Creates a LatLng object representing a geographic point.
   * Note the ordering of latitude and longitude.
   * @param lat Latitude is specified in degrees within the range [-90, 90].
   * @param lng Longitude is specified in degrees within the range [-180, 180].
   * @param noWrap Set noWrap to true to enable values outside of this range.
   */
  constructor(lat: number, lng: number, noWrap?: boolean);
  /** Comparison function. */
  equals(other: LatLng): boolean;
  /** Returns the latitude in degrees. */
  lat(): number;
  /** Returns the longitude in degrees. */
  lng(): number;
  /** Converts to string representation. */
  toString(): string;
  /** Returns a string of the form "lat,lng". We round the lat/lng values to 6 decimal places by default. */
  toUrlValue(precision?: number): string;
}

export type LatLngLiteral = { lat: number; lng: number };

export class LatLngBounds {
  constructor(sw?: LatLng, ne?: LatLng);
  contains(latLng: LatLng): boolean;
  equals(other: LatLngBounds): boolean;
  extend(point: LatLng): LatLngBounds;
  getCenter(): LatLng;
  getNorthEast(): LatLng;
  getSouthWest(): LatLng;
  intersects(other: LatLngBounds): boolean;
  isEmpty(): boolean;
  toSpan(): LatLng;
  toString(): string;
  toUrlValue(precision?: number): string;
  union(other: LatLngBounds): LatLngBounds;
}
// ...
export class MVCObject {
  constructor();
  addListener(
    eventName: string,
    handler: (...args: any[]) => void
  ): MapsEventListener;
  bindTo(
    key: string,
    target: MVCObject,
    targetKey?: string,
    noNotify?: boolean
  ): void;
  changed(key: string): void;
  get(key: string): any;
  notify(key: string): void;
  set(key: string, value: any): void;
  setValues(values: any): void;
  unbind(key: string): void;
  unbindAll(): void;
}
// ...
export class Autocomplete extends MVCObject {
  constructor(inputField: HTMLInputElement, opts?: AutocompleteOptions);
  getBounds(): LatLngBounds;
  getPlace(): PlaceResult;
  setBounds(bounds: LatLngBounds): void;
  setComponentRestrictions(restrictions: ComponentRestrictions): void;
  setTypes(types: string[]): void;
}

export interface AutocompleteOptions {
  bounds?: LatLngBounds;
  componentRestrictions?: ComponentRestrictions;
  types?: string[];
}
// ...
export interface ComponentRestrictions {
  country: string;
}

export interface PlaceAspectRating {
  rating: number;
  type: string;
}

export interface PlaceDetailsRequest {
  placeId: string;
}

export interface PlaceGeometry {
  location: LatLng;
  viewport: LatLngBounds;
}

export interface PlacePhoto {
  height: number;
  html_attributions: string[];
  width: number;
  getUrl(opts: PhotoOptions): string;
}

export interface PhotoOptions {
  maxHeight?: number;
  maxWidth?: number;
}

export interface PlaceResult {
  address_components: GeocoderAddressComponent[];
  aspects: PlaceAspectRating[];
  formatted_address: string;
  formatted_phone_number: string;
  geometry: PlaceGeometry;
  html_attributions: string[];
  icon: string;
  international_phone_number: string;
  name: string;
  permanently_closed: boolean;
  photos: PlacePhoto[];
  place_id: string;
  price_level: number;
  rating: number;
  reviews: PlaceReview[];
  types: string[];
  url: string;
  vicinity: string;
  website: string;
}

export interface PlaceReview {
  aspects: PlaceAspectRating[];
  author_name: string;
  author_url: string;
  language: string;
  text: string;
}
// ...
  export class MapsEventListener {}